+++ test/Makefile
@@ -1,0 +1,3 @@
+EVERQUIC_HOME=$(realpath ..)
+
+include $(EVERQUIC_HOME)/Makefile.include
